feat: add RecordKey strict parser conforming to LexiconStringFormat
